+++ odoo/local-src/mtsmte_reports/models/helpers.py
@@ -15,6 +15,21 @@ class Helpers(models.AbstractModel):
         'nb_shocks',
     )
 
+    def _get_conformity_options(self, lang):
+        """Returns translated options for conformity
+
+        Due to mysterious reasons conformity selection was not
+        getting translated in the project analysis report. After
+        headbanging it with Simone for sometime it was decided to
+        hack it through the roof.
+        :param lang: language to translate the field to
+        :returns: dictionary with translated values
+        """
+        model = self.env["product.substance.measure"].with_context(lang=lang)
+        return dict(
+            model.fields_get(['conformity'])['conformity']['selection']
+        )
+
     @api.model
     def render_html(self, docids, data=None):
         report_obj = self.env['report']
@@ -23,16 +38,19 @@ class Helpers(models.AbstractModel):
         if docids:
             docs = self.env[report.model].browse(docids)
             tasks = []
+            lang = self.env.lang
         else:
             docs = self.env[report.model].browse(data['project_id'])
             tasks = self.env['project.task'].browse(data['task_ids'])
+            lang = data.get('forced_lang') or self.env.lang
         docargs = {
             'doc_ids': docids,
             'doc_model': report.model,
             'docs': docs,
             'tasks': tasks,
             'report': self,
-            'forced_lang': data.get('forced_lang'),
+            'forced_lang': lang,
+            'conformity_options': self._get_conformity_options(lang),
         }
         return report_obj.render(report_name, docargs)